% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{getLatestVdjdb}
\alias{getLatestVdjdb}
\alias{annotateTCR}
\title{download the lastest vdjdb-db}
\usage{
getLatestVdjdb()

annotateTCR(btcr, latest = FALSE)
}
\arguments{
\item{btcr}{A BTCR object.}

\item{latest}{Use the latest vdjdb data (auto download).}
}
\value{
with attribute "annotation" added to btcr.
}
\description{
download the lastest vdjdb-db

annotate TCR with vdjdb
}
