% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{assignStatus}
\alias{assignStatus}
\title{assign status: clonal or not}
\usage{
assignStatus(btcr, idcol = "clonotype_id", scrna = NULL,
  group.by = NULL)
}
\arguments{
\item{idcol}{Column name used as the clonotype id.}

\item{scrna}{The Single Cell RNAseq Seurat object corresponded to the BCR or TCR.}

\item{group.by}{The meta.data column in scrna used to group the cells.}
}
\value{
A new BTCR object with "status" and "status2"(if scrna and group.by provided).
}
\description{
assign status: clonal or not
}
\details{
"status" is used to indicate whether a cell is clonal or not, "status2" is used to indicate their groups relationship.
  For each clonotypes, if there are at least two cells with it, these cells would be clonals.
  For each clonals, if they are from different cell group, they are cross-cluster; otherwise, they are within-cluster.
}
