% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{barPointError}
\alias{barPointError}
\title{barplot with points and error bars based on mulitple samples data.}
\usage{
barPointError(data, x, y, fill, point = TRUE, errbar = TRUE)
}
\arguments{
\item{data}{The data frame to plot.}

\item{x}{The column used as X axis.}

\item{y}{THe column used as Y axis.}

\item{fill}{THe column used as groups.}

\item{point}{Whether to plot points.}

\item{errbar}{Wheter to plot error bars.}
}
\value{
ggplot.
}
\description{
barplot with points and error bars based on mulitple samples data.
}
