% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{calcOverlap}
\alias{calcOverlap}
\title{calculate the clonotype overlaps in each group}
\usage{
calcOverlap(btcr, scrna, groupby, splitby = NULL,
  id = c("final_clonotype_id", "clonotype_id", "clonotype_newid"))
}
\arguments{
\item{btcr}{A BTCR object.}

\item{scrna}{scrna The Single Cell RNAseq Seurat object corresponded to the BCR or TCR.}

\item{groupby}{Overlaps groups. The meta.data column used to group cells/barcodes, always be clusters.}

\item{splitby}{Sample groups. Overlaps won't be calculated cross groups. The meta.data column used to group cells/barcodes, always be sample groups.}

\item{id}{Which clonotype id used to calculate overlaps.}
}
\value{
A BTCR object with "overlaps" and "id" attributes.
}
\description{
calculate the clonotype overlaps in each group
}
