% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{filterBtcr}
\alias{filterBtcr}
\title{filter data}
\usage{
filterBtcr(btcr, scrna = NULL)
}
\arguments{
\item{btcr}{The BTCR object.}

\item{scrna}{The Single Cell RNAseq Seurat object corresponded to the BCR or TCR. Used for filtering the barcodes.}
}
\value{
A BTCR filtered by non-productive chains and invalid barcodes.
}
\description{
Remove the Non-productive chains. If scrna is provided, cells not in scrna would be filtered either.
}
