% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{loadSingle}
\alias{loadSingle}
\alias{loadMultiple}
\title{load data}
\usage{
loadSingle(path, sample, CR = c("BCR", "TCR"))

loadMultiple(paths, groups = NULL, CR = c("BCR", "TCR"))
}
\arguments{
\item{path}{Path to the Cellranger VDJ outs directory.}

\item{sample}{Name for this sample.}

\item{CR}{TCR or BCR.}

\item{paths}{Named paths to Cellranger VDJ outs directory.}

\item{groups}{Groups corresponded to each sample.}
}
\value{
A BTCR list contains "clonotypes", "contigs" and "merge" data.
}
\description{
load Cellranger VDJ data, including BCR and TCR.
}
\details{
The "merge" element in BTCR list object is a combination of "clonotypes" and "contigs". In "merge", each gene isoforms are pasted togother so that every cell has just one line.
}
