% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotAnnoDim}
\alias{plotAnnoDim}
\title{plot the antigen species}
\usage{
plotAnnoDim(btcr, scrna, dtype = c("TRA", "TRB"), ...)
}
\arguments{
\item{btcr}{The BTCR object annotated by annotateTCR}

\item{scrna}{The Single Cell RNAseq Seurat object corresponded to the BCR or TCR.}

\item{dtype}{"TRA" or "TRB". This function supports TCR only.}

\item{...}{Extra params(except "group.by") passed to \code{Seurat::DimPlot}.}
}
\value{
\code{Seurat::DimPlot} returns.
}
\description{
plot the antigen species
}
