% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotCdr3}
\alias{plotCdr3}
\title{barplot of top n cdr3 sequences usage in each group.}
\usage{
plotCdr3(btcr, group, dtype = c("percentage", "count"), topn = 10,
  ncol = 1, combine = TRUE)
}
\arguments{
\item{btcr}{A BTCR object.}

\item{group}{Column in btcr$contigs to group cdr3.}

\item{dtype}{Plot usage counts or percentages.}

\item{topn}{Top N cdr3 to plot. Default: 10.}

\item{ncol}{Ncols in plot layout.}

\item{combine}{Combine each plots or return plot list.}
}
\value{
patchwork wrapped plots or ggplot list.
}
\description{
barplot of top n cdr3 sequences usage in each group.
}
