% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotCellCR}
\alias{plotCellCR}
\title{barplot of cells with TCR or BCR percentages in grouped T cells or B cells}
\usage{
plotCellCR(btcr, scrna, group)
}
\arguments{
\item{btcr}{A BTCR object.}

\item{scrna}{The Single Cell RNAseq Seurat object corresponded to the BCR or TCR.}

\item{group}{The meta.data column in scrna to group cells.}
}
\value{
A ggplot object.
}
\description{
barplot of cells with TCR or BCR percentages in grouped T cells or B cells
}
