% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotCloneDim}
\alias{plotCloneDim}
\title{dimension plots of clonotypes}
\usage{
plotCloneDim(btcr, scrna, colorby = c("CR", "cdr3s_aa", "cdr3s_nt"),
  topn = 5, highlight = NULL, ...)
}
\arguments{
\item{btcr}{A BTCR object.}

\item{scrna}{The Single Cell RNAseq Seurat object corresponded to the BCR or TCR.}

\item{colorby}{can be "CR", "cdr3s_aa", "cdr3s_nt". If "CR", only plot whether a cell has a TCR or BCR. If "cdr3s", use topn or highlight to limit the clonotypes.}

\item{topn}{Top N barcode frequency cdr3s to plot. Will be ignored if highlight is provided.}

\item{highlight}{Which cdr3s in colorby to plot.}

\item{...}{Arguments for \code{\link[Seurat::DimPlot]{Seurat::DimPlot}}.}
}
\value{
Seurat::DimPlot returns.
}
\description{
dimension plots of clonotypes
}
