% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotCloneState}
\alias{plotCloneState}
\title{barplot of distributions of clone status in grouped T cells or B cells.}
\usage{
plotCloneState(btcr, scrna, group.by, split.by = NULL,
  status = c("status", "status2"))
}
\arguments{
\item{btcr}{A BTCR object.}

\item{scrna}{The Single Cell RNAseq Seurat object corresponded to the BCR or TCR.}

\item{group.by}{The meta.data column in scrna to group cells, like clusters.}

\item{split.by}{The meta.data column in scrna to group cells, like sample groups.}

\item{status}{The clonal status used.}
}
\value{
ggplot.
}
\description{
barplot of distributions of clone status in grouped T cells or B cells.
}
