% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotGeneBar}
\alias{plotGeneBar}
\title{barplot of each VDJ gene usages}
\usage{
plotGeneBar(btcr, group, sample = "sample", gene = c("v", "j", "c"),
  dtype = c("percentage", "count"), isoform = TRUE, combine = TRUE,
  rotate = TRUE, ncol = 1)
}
\arguments{
\item{btcr}{A BTCR object.}

\item{group}{Column in btcr$contigs to group each VDJ genes, always be sample or group.}

\item{sample}{The sample column in btcr$contigs.}

\item{gene}{Plot v_gene, j_gene or c_gene.}

\item{dtype}{Plot barcode usage counts or percentages.}

\item{isoform}{If set TRUE, plot the gene isoforms like TRAV-xxx. Or just plot the main gene like TRA, TRB (isoforms are summed.).}

\item{combine}{Wrap plots togother. Or return the plot list.}

\item{rotate}{Rotate the x labels.}

\item{ncol}{When plot isoform, plot ncol in the layout.}
}
\value{
A ggplot or ggplot list or patchwork wrapped plot.
}
\description{
barplot of each VDJ gene usages
}
