% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotSharedCloneDim}
\alias{plotSharedCloneDim}
\title{dimension plots of shared or unique clonotypes in each groups.}
\usage{
plotSharedCloneDim(btcr, scrna, values, dtype = c("share", "unique"),
  ncol = 1, ...)
}
\arguments{
\item{btcr}{A overlaps-calculated BTCR object. see \code{\link{calcOverlap}}.}

\item{scrna}{The Single Cell RNAseq Seurat object corresponded to the BCR or TCR.}

\item{values}{The overlapped groups to plot, at least two values in \code{attributes(btcr)$group}.}

\item{dtype}{Plot the shared clonotypes or the unique clonotypes.}

\item{ncol}{Ncols in plot layout.}

\item{...}{Arguments for \code{\link[Seurat::DimPlot]{Seurat::DimPlot}}.}
}
\value{
Seurat::DimPlot returns.
}
\description{
dimension plots of shared or unique clonotypes in each groups.
}
