% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotVJpair}
\alias{plotVJpair}
\title{heatmap of v-j pairs usage}
\usage{
plotVJpair(btcr, group, dtype = c("count", "percentage"), topn = 10,
  ncol = 1, position = c("right", "left", "bottom", "top"))
}
\arguments{
\item{btcr}{A BTCR object.}

\item{group}{COlumn in btcr$contigs to group genes.}

\item{dtype}{Plot usage counts or percentages.}

\item{topn}{Just plot the v-j pairs with top N usage. Default: 10.}

\item{ncol}{Ncols in plot layout.}

\item{position}{The position of group labels in plot.}
}
\value{
ggplot
}
\description{
heatmap of v-j pairs usage
}
