% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{refreshId}
\alias{refreshId}
\title{create the final clonotype id}
\usage{
refreshId(btcr, by = c("cdr3s_nt", "cdr3s_aa"))
}
\arguments{
\item{btcr}{The BTCR object.}

\item{by}{use the "cdr3s_nt" or "cdr3s_aa" to create id. Default is nt.}
}
\value{
A new BTCR object with "final_clonotype_id" column.
}
\description{
create the final clonotype id
}
