% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{saveBtcr}
\alias{saveBtcr}
\title{save the BTCR data to a csv file}
\usage{
saveBtcr(btcr, outfile)
}
\arguments{
\item{btcr}{A BTCR object.}

\item{outfile}{The path to file to write in. It should be with a "csv" suffix.}
}
\description{
save the BTCR data to a csv file
}
