% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{saveBtcr}
\alias{saveBtcr}
\title{save the BTCR data to a csv file}
\usage{
saveBtcr(btcr, outpath = ".")
}
\arguments{
\item{btcr}{A BTCR object.}

\item{outpath}{The path to save data.}
}
\description{
save the BTCR data to a csv file
}
